fix: move status to own file
